+++ src/onegov.gis/HISTORY.rst
@@ -1,6 +1,9 @@
 Changelog
 ---------
 
+- Acquires the default view parameters from the body.
+  [href]
+
 0.1.0 (2016-04-05)
 ~~~~~~~~~~~~~~~~~~~